+++ scripts/lint-allow-test-rule-refs.allowlist.json
@@ -17,28 +17,17 @@
   "tests/bug-131-release-tarball-smoke-explicit-home.test.cjs :: integration-test-input",
   "tests/bug-211-launcher-home-fallback.test.cjs :: structural/behavioral regression for the ~/.claude fallback arm in",
   "tests/bug-2136-sh-hook-version.test.cjs :: structural-regression-guard",
-  "tests/bug-214-phase-researcher-write-truncation-contract.test.cjs :: source-text-is-the-product",
-  "tests/bug-214-writer-agents-write-truncation-contract.test.cjs :: source-text-is-the-product",
-  "tests/bug-2346-agent-read-loop-guards.test.cjs :: source-text-is-the-product",
-  "tests/bug-2419-project-researcher-agent.test.cjs :: source-text-is-the-product",
-  "tests/bug-2421-planner-grep-gate-hygiene.test.cjs :: source-text-is-the-product",
   "tests/bug-2543-gsd-slash-namespace.test.cjs :: structural-regression-guard",
   "tests/bug-2559-stale-search-year.test.cjs :: source-text-is-the-product",
-  "tests/bug-2686-review-fix-worktree.test.cjs :: source-text-is-the-product",
   "tests/bug-2772-gitmodules-path-intersection.test.cjs :: source-text-is-the-product",
   "tests/bug-2808-skill-hyphen-name.test.cjs :: source-text-is-the-product",
   "tests/bug-2839-review-fix-transactional-cleanup.test.cjs :: source-text-is-the-product",
-  "tests/bug-2990-code-fixer-worktree-branch.test.cjs :: source-text-is-the-product",
-  "tests/bug-3097-3099-executor-worktree-path-safety.test.cjs :: reads markdown product files (gsd-executor.md, worktree-path-safety.md) to verify structural protocol — not source-grep",
-  "tests/bug-3290-intel-updater-layout-block.test.cjs :: source-text-is-the-product — agents/gsd-intel-updater.md IS",
   "tests/bug-33-settings-model-profile-adaptive.test.cjs :: source-text-is-the-product",
   "tests/bug-3384-secondary-defects.test.cjs :: source-text-is-the-product",
-  "tests/bug-3430-planner-phase-contract.test.cjs :: source-text-is-the-product",
   "tests/bug-3446-resume-continue-here-discovery.test.cjs :: source-text-is-the-product",
   "tests/bug-3491-nested-git-worktree.test.cjs :: source-text-is-the-product",
   "tests/bug-3523-cjs-loadconfig-branching-strategy-warning.test.cjs :: validates runtime CLI stdout/stderr warning behavior, not source grep",
   "tests/bug-3542-executor-git-stash-prohibition.test.cjs :: source-text-is-the-product",
-  "tests/bug-3605-stale-research-insert-phase-agent-refs.test.cjs :: source-text-is-the-product",
   "tests/bug-3677-agent-colon-namespace-leak.test.cjs :: source-text-is-the-product",
   "tests/bug-3678-executor-commit-docs-respect.test.cjs :: source-text-is-the-product",
   "tests/bug-3683-command-colon-namespace-leak.test.cjs :: source-text-is-the-product",
@@ -46,7 +35,6 @@
   "tests/bug-3689-resume-glob-nomatch.test.cjs :: source-text-is-the-product",
   "tests/bug-3810-no-gsd-sdk-runtime-refs.test.cjs :: source-text-is-the-product",
   "tests/bug-444-resolver-local-claude-install.test.cjs :: structural/behavioral regression for the repo-local .claude/ install",
-  "tests/bug-571-doc-writer-fix-mode-edit-only.test.cjs :: source-text-is-the-product",
   "tests/bug-619-codebase-drift-gate-shim.test.cjs :: source-text-is-the-product",
   "tests/bug-622-graphify-optional-graph-html.test.cjs :: source-text-is-the-product",
   "tests/bug-630-wave-cleanup-orchestrator-root.test.cjs :: source-text-is-the-product",
@@ -91,7 +79,6 @@
   "tests/edge-probe-spec-phase-contract.test.cjs :: runtime-contract-is-the-product — spec-phase.md Step 5.5 is the deployed workflow runtime contract under assertion",
   "tests/edit-phase.test.cjs :: source-text-is-the-product",
   "tests/enh-2380-sync-skills.test.cjs :: source-text-is-the-product",
-  "tests/enh-2500-codebase-mapper-arch-rich-format.test.cjs :: source-text-is-the-product",
   "tests/enh-2789-description-budget.test.cjs :: source-text-is-the-product",
   "tests/enh-2790-skill-consolidation.test.cjs :: source-text-is-the-product",
   "tests/enh-48-cwd-drift-guard-e2e.test.cjs :: integration-test-input",
