+++ scripts/lint-regression-test-names.allowlist.json
@@ -6,37 +6,24 @@
   "bug-21-state-md-template-frontmatter.test.cjs",
   "bug-211-launcher-home-fallback.test.cjs",
   "bug-2136-sh-hook-version.test.cjs",
-  "bug-214-phase-researcher-write-truncation-contract.test.cjs",
-  "bug-214-writer-agents-write-truncation-contract.test.cjs",
   "bug-2344-read-guard-claudecode-env.test.cjs",
-  "bug-2346-agent-read-loop-guards.test.cjs",
-  "bug-2351-intel-kilo-layout.test.cjs",
-  "bug-2419-project-researcher-agent.test.cjs",
-  "bug-2421-planner-grep-gate-hygiene.test.cjs",
   "bug-2451-context-monitor-over-report.test.cjs",
   "bug-2520-read-guard-hook-subprocess-env.test.cjs",
   "bug-2543-gsd-slash-namespace.test.cjs",
   "bug-2559-stale-search-year.test.cjs",
   "bug-260-worktree-path-guard.test.cjs",
   "bug-261-worktree-force-add-guard.test.cjs",
-  "bug-2686-review-fix-worktree.test.cjs",
   "bug-2772-gitmodules-path-intersection.test.cjs",
   "bug-2808-skill-hyphen-name.test.cjs",
   "bug-2839-review-fix-transactional-cleanup.test.cjs",
   "bug-2866-codex-strip-no-trailing-newline.test.cjs",
   "bug-2876-skill-frontmatter-quote.test.cjs",
   "bug-2916-handle-branching-default-base.test.cjs",
-  "bug-2990-code-fixer-worktree-branch.test.cjs",
   "bug-2995-post-install-script-paths.test.cjs",
   "bug-3019-help-passthrough.test.cjs",
   "bug-3054-stale-gsd-next-references.test.cjs",
-  "bug-3087-planner-directive-language.test.cjs",
-  "bug-3097-3099-executor-worktree-path-safety.test.cjs",
-  "bug-3290-intel-updater-layout-block.test.cjs",
   "bug-33-settings-model-profile-adaptive.test.cjs",
-  "bug-3321-verifier-runs-probes.test.cjs",
   "bug-3384-secondary-defects.test.cjs",
-  "bug-3430-planner-phase-contract.test.cjs",
   "bug-3442-shim-projection-drift-guard.test.cjs",
   "bug-3446-resume-continue-here-discovery.test.cjs",
   "bug-3491-nested-git-worktree.test.cjs",
@@ -44,7 +31,6 @@
   "bug-3523-cjs-loadconfig-branching-strategy-warning.test.cjs",
   "bug-3542-executor-git-stash-prohibition.test.cjs",
   "bug-3588-npm-audit-clean.test.cjs",
-  "bug-3605-stale-research-insert-phase-agent-refs.test.cjs",
   "bug-3668-workflow-runtime-resolution.test.cjs",
   "bug-3677-agent-colon-namespace-leak.test.cjs",
   "bug-3678-executor-commit-docs-respect.test.cjs",
@@ -53,7 +39,6 @@
   "bug-3689-resume-glob-nomatch.test.cjs",
   "bug-3810-no-gsd-sdk-runtime-refs.test.cjs",
   "bug-444-resolver-local-claude-install.test.cjs",
-  "bug-571-doc-writer-fix-mode-edit-only.test.cjs",
   "bug-619-codebase-drift-gate-shim.test.cjs",
   "bug-622-graphify-optional-graph-html.test.cjs",
   "bug-630-wave-cleanup-orchestrator-root.test.cjs",
